requirements.txt
python-telegram-bot[webhooks]==21.4
flask==3.0.0
flask-cors==4.0.0
